+++ CHANGELOG.md
@@ -18,3 +18,10 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.7] - 2024-11-15
 ### Added
 - File text content inclusion functionality with automatic content synchronization
+
+## [0.0.9] - 2024-11-24
+### Removed
+- File text content inclusion functionality with automatic content synchronization. Will be available as part of a separate extention `smart-portals`
+
+### Added
+- Including VSCode diagnostics output of selected files, as part of copied stream